Squashing more compiler warnings

+++ SpendingTracker/Client/Pages/Subcategories/SubcategoryDetail.razor.cs
@@ -8,10 +8,9 @@ namespace SpendingTracker.Client.Pages.Subcategories;
 public partial class SubcategoryDetail
 {
     [Parameter] public string? Id { get; set; }
-    [Inject] private ISubcategoriesService SubcategoriesService { get; set; }
-    [Inject] private NavigationManager NavigationManager { get; set; }
-    [Inject] private ISnackbar Snackbar { get; set; }
-    private bool IsLoading { get; set; } = false;
+    [Inject] private ISubcategoriesService SubcategoriesService { get; set; } = default!;
+    [Inject] private NavigationManager NavigationManager { get; set; } = default!;
+    [Inject] private ISnackbar Snackbar { get; set; } = default!;
     private bool DialogVisible { get; set; } = false;
     private void ToggleDialog() => DialogVisible = !DialogVisible;
 
@@ -19,20 +18,23 @@ public partial class SubcategoryDetail
 
     protected override async Task OnInitializedAsync()
     {
-        IsLoading = true;
-        Subcategory = await SubcategoriesService.GetSubcategory(int.Parse(Id));
-        IsLoading = false;
+        if (Id != null)
+        {
+            Subcategory = await SubcategoriesService.GetSubcategory(int.Parse(Id));
+        }
     }   
     private async Task DeleteSubcategory()
     {
         try
         {
-            await SubcategoriesService.DeleteSubcategory(int.Parse(Id));
+            if (Id != null)
+            {
+                await SubcategoriesService.DeleteSubcategory(int.Parse(Id));
+                ToggleDialog();
+                Snackbar.Add($"Subcategory removed", Severity.Success);
                 
-            ToggleDialog();
-            Snackbar.Add($"Subcategory removed", Severity.Success);
-                
-            NavigationManager.NavigateTo("/categories", false);
+                NavigationManager.NavigateTo("/categories", false);
+            }
         }
         catch
         {
